User newsfeed content pulled through api calls and displayed in a makeshift textview.
Twitter object is used to obtain request and access tokens.  The same twitter object is used in conjunction with the access token and access token secret to pull user newsfeed data by interfacing with the twitter servers through the twitter4j apis. The makeshift textview is updated with data after the user has clicked the log in button and signed in.